Made the user page dynamic

+++ app/api/tasks/[id]/route.ts
@@ -1,22 +1,35 @@
 import prisma from "@/prisma/client";
 import { NextRequest, NextResponse } from "next/server";
+import { getServerSession } from "next-auth";
+import { authOptions } from "../../auth/[...nextauth]/options";
 
 export async function PATCH(
   request: NextRequest,
   { params }: { params: { id: string } }
 ) {
   const body = await request.json();
+  const session = await getServerSession(authOptions);
 
+  if (!session || !session.user?.id) {
+    return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+  }
+
+  const taskId = parseInt(params.id);
   const task = await prisma.task.findUnique({
-    where: { id: parseInt(params.id) },
+    where: { id: taskId },
   });
 
   if (!task) {
-    return NextResponse.json({ error: "Invalid task" }, { status: 404 });
+    return NextResponse.json({ error: "Task not found" }, { status: 404 });
+  }
+
+  // Ensure the task belongs to the current user
+  if (task.userId !== session.user.id) {
+    return NextResponse.json({ error: "Forbidden" }, { status: 403 });
   }
 
   const updatedTask = await prisma.task.update({
-    where: { id: task.id },
+    where: { id: taskId },
     data: {
       status: body.status,
       completedAt: body.status === "COMPLETED" ? new Date() : null,
@@ -30,6 +43,12 @@ export async function DELETE(
   request: NextRequest,
   { params }: { params: { id: string } }
 ) {
+  const session = await getServerSession(authOptions); // Adjust based on your auth setup
+
+  if (!session || !session.user?.id) {
+    return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+  }
+
   const taskId = parseInt(params.id);
 
   if (isNaN(taskId)) {
@@ -41,7 +60,12 @@ export async function DELETE(
   });
 
   if (!task) {
-    return NextResponse.json({ error: "Invalid task" }, { status: 404 });
+    return NextResponse.json({ error: "Task not found" }, { status: 404 });
+  }
+
+  // Ensure the task belongs to the current user
+  if (task.userId !== session.user?.id) {
+    return NextResponse.json({ error: "Forbidden" }, { status: 403 });
   }
 
   await prisma.task.delete({
